+++ docs/tutorials/python/background-check/code/backgroundcheck_boilerplate/activities/ssntraceactivity.py
@@ -1,0 +1,8 @@
+# @@@SNIPSTART python-project-setup-chapter-backgroundcheck-activity
+from temporalio import activity
+
+@activity.defn
+async def ssn_trace_activity(ssn) -> str:
+    return "pass"
+
+# @@@SNIPEND
