+++ docs/tutorials/python/background-check/code/backgroundcheck_boilerplate/cloud_worker/main.py
@@ -1,0 +1,39 @@
+# @@@SNIPSTART python-project-setup-chapter-cloud-worker
+import asyncio
+import os
+
+from temporalio.client import Client, TLSConfig
+from temporalio.worker import Worker
+
+from activities.ssntraceactivity_dacx import ssn_trace_activity
+from workflows.backgroundcheck_dacx import BackgroundCheck
+
+async def main():
+    with open(os.getenv("TEMPORAL_MTLS_TLS_CERT"), "rb") as f:
+        client_cert = f.read()
+
+    with open(os.getenv("TEMPORAL_MTLS_TLS_KEY"), "rb") as f:
+        client_key = f.read()
+
+    client = await Client.connect(
+        os.getenv("TEMPORAL_HOST_URL"),
+        namespace=os.getenv("TEMPORAL_NAMESPACE"),
+        tls=TLSConfig(
+            client_cert=client_cert,
+            client_private_key=client_key,
+        ),
+    )
+
+    worker = Worker(
+        client,
+        task_queue="backgroundcheck-boilerplate-task-queue",
+        workflows=[BackgroundCheck],
+        activities=[ssn_trace_activity],
+    )
+    await worker.run()
+
+
+if __name__ == "__main__":
+    asyncio.run(main())
+
+# @@@SNIPEND
