+++ docs/tutorials/python/background-check/code/backgroundcheck_boilerplate/workflows/backgroundcheck.py
@@ -1,23 +1,19 @@
-# @@@SNIPSTART python-durability-chapter-backgroundcheck-workflow
-import asyncio
+# @@@SNIPSTART python-project-setup-chapter-backgroundcheck-workflow
 from datetime import timedelta
 
 from temporalio import workflow
 
 with workflow.unsafe.imports_passed_through():
-    from ssntraceactivity import ssn_trace_activity
+    from activities.ssntraceactivity_dacx import ssn_trace_activity
 
-@workflow.defn()
+@workflow.defn
 class BackgroundCheck:
     @workflow.run
     async def run(self, ssn: str) -> str:
-        random_number = workflow.random().randint(1, 100)
-        if random_number < 50:
-            await asyncio.sleep(60)
-            workflow.logger.info("Sleeping for 60 seconds")
         return await workflow.execute_activity(
             ssn_trace_activity,
             ssn,
             schedule_to_close_timeout=timedelta(seconds=5),
         )
+
 # @@@SNIPEND
