+++ docs/tutorials/python/background-check/index.mdx
@@ -14,7 +14,13 @@ keywords:
 ---
 
 Welcome to the Temporal Python SDK Background Check tutorial!
-Learn Temporal and its features while building a Background Check application.
+
+In this tutorial, you will build a Background Check application using the Temporal Python SDK and explore how to do the following:
+
+- **Construct a new Temporal Application project.**
+- **Develop applications for Durable Execution.**
+
+The tutorial is laid out in the following chapters:
 
 - [Introduction](/tutorials/python/background-check/introduction.mdx)
 - [Project setup](/tutorials/java/background-check/project-setup.mdx)
