+++ docs/tutorials/python/background-check/introduction.mdx
@@ -13,7 +13,7 @@ tags:
 ---
 
 
-Welcome to Temporal Python SDK developer's guide!
+Welcome to Temporal Python SDK Background Check tutorial!
 
 :::info Temporal Python SDK API reference
 
@@ -103,7 +103,7 @@ For complex and large-scale use cases, having some experience with the following
 
 ### Where can I find code samples? {#samples}
 
-Code samples are integrated into this developer’s guide.
+Code samples are integrated into this tutorial.
 You can find those code samples in the [temporalio/documentation-samples-python](https://github.com/temporalio/documentation-samples-python) repository on GitHub.
 
 Additional Python code samples are in the [temporalio/samples-python](https://github.com/temporalio/samples-python) repository on GitHub.
